Revert "Tidy Up"
This reverts commit f2cff50ff8f8fa1471af3d595ad666ace96e1ec7.

+++ UnitTests/SeleniumDriver/SeleniumDriverTestSteps.cs
@@ -1,30 +1,37 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
+﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using TeamControlium.Utilities;
+using System.Text;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using TechTalk.SpecFlow;
+using System.Threading;
+using TeamControlium.Utilities;
 
 namespace TeamControlium.Controlium
 {
     [Binding]
     public sealed class SeleniumDriverTestSteps
     {
+ 
         [Given(@"There are no processes running named ""(.*)""")]
         public void GivenThereAreNoProcessesRunningNamed(string processNameToKill)
         {
             Process[] processes = Process.GetProcessesByName(processNameToKill);
-            foreach (var process in processes)
+            if (processes.Count() > 1)
             {
-                try
+                foreach (var process in processes)
                 {
-                    process.Kill();
-                    if (!process.WaitForExit(10000))
+                    try
                     {
-                        Assert.Inconclusive($"Killing {processNameToKill}] (ID = {process.Id}) did not work within 10 seconds");
+                        process.Kill();
+                        if (!process.WaitForExit(10000))
+                        {
+                            Assert.Inconclusive($"Killing {processNameToKill}] (ID = {process.Id}) did not work within 10 seconds");
+                        }
                     }
+                    catch (Exception ex) { }
                 }
-                catch { }
             }
         }
 
@@ -38,15 +45,18 @@ namespace TeamControlium.Controlium
         [When(@"I instantiate SeleniumDriver")]
         public void WhenIInstantiateSeleniumDriver()
         {
-            ScenarioContext.Current.Add("SeleniumDriver", new SeleniumDriver());
+            SeleniumDriver sDriver = new SeleniumDriver();
+            
+            ScenarioContext.Current.Add("SeleniumDriver", sDriver);
         }
 
         [Given(@"I browse to ""(.*)""")]
         [When(@"I browse to ""(.*)""")]
         public void WhenIBrowseTo(string urlToBrowseTo)
         {
-            ScenarioContext.Current.Get<SeleniumDriver>("SeleniumDriver").GotoURL(urlToBrowseTo);
+            ((SeleniumDriver)ScenarioContext.Current["SeleniumDriver"]).GotoURL(urlToBrowseTo);
         }
+
 
         [Then(@"a process exists named ""(.*)""")]
         public void ThenAProcessExistsNamed(string processname)
@@ -57,7 +67,8 @@ namespace TeamControlium.Controlium
         [Then(@"I can read the page title ""(.*)""")]
         public void ThenICanReadThePageTitle(string expectedPageTitle)
         {
-            Assert.AreEqual(expectedPageTitle, ScenarioContext.Current.Get<SeleniumDriver>("SeleniumDriver").PageTitle);
+            string pageTitle = ((SeleniumDriver)ScenarioContext.Current["SeleniumDriver"]).PageTitle;
+            Assert.AreEqual(expectedPageTitle, pageTitle);
         }
 
         [Given(@"I use FindElement to locate an element using XPath ""(.*)""")]
@@ -67,7 +78,7 @@ namespace TeamControlium.Controlium
             Element foundElement;
             try
             {
-                foundElement = ScenarioContext.Current.Get<SeleniumDriver>("SeleniumDriver").FindElement(new ObjectMappingDetails(xPath, "The XPath"));
+                foundElement = ((SeleniumDriver)ScenarioContext.Current["SeleniumDriver"]).FindElement(new ObjectMappingDetails(xPath, "The XPath"));
                 ScenarioContext.Current.Add("FoundElement", foundElement);
             }
             catch (Exception ex)
@@ -76,10 +87,14 @@ namespace TeamControlium.Controlium
             }
         }
 
+
         [Then(@"a valid element is found")]
         public void ThenAValidElementIsFound()
         {
             Assert.IsNotNull(ScenarioContext.Current["FoundElement"], "Found element must not be null");
         }
+
+
+
     }
 }
